enterprise-genai-suite/mvp2-agentic-ai-interior-stylist/requirements.txt
openai
langchain
python-dotenv
streamlit
